pyproject.toml
[tool.poetry]
name = "sesame"
version = "1.1.3"
description = "A helper program to seal the secrets in the bulk"
authors = ["Michael Tesar <michael.tesar@rossum.ai>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.282"

[tool.poetry.scripts]
sesame = "sesame.sesame:main_cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
